From: Raphael Collado <raphael.collado@st.com> Bug fix: NDEF not detected.
The NFC Forum Type 4 specifications mention that values 0x80 to 0xFE for the NDEF file
write access condition byte are proprietary. In the current library those values are considered
incorrect. Being considered incorrect when the Tag value is different from 0x00 or 0xFF, the stack
returns an error leading to a non detection of an NDEF message.
Here below the logcat on An Android phone :
12-17 03:37:01.157  3321  3628 E BrcmNfcNfa: rw_t4t_validate_cc_file (): Write Access (0x80) is invalid
12-17 03:37:01.157  3321  3628 E BrcmNfcNfa: rw_t4t_handle_error (): status:0xE7, sw1:0x00, sw2:0x00, state:0x2

According Type 4 specifications 0x80 is valid.